+++ test/repositories/ChainContractRepository.test.ts
@@ -2,7 +2,7 @@ import 'reflect-metadata';
 import { expect } from 'chai';
 import { ChainContractRepository } from '../../src/repositories/ChainContractRepository';
 import { BlockchainRepository } from '../../src/repositories/BlockchainRepository';
-import { ForeignChainsIds, NonEvmChainsIds } from '../../src/types/ChainsIds';
+import { ChainsIds, ForeignChainsIds, NonEvmChainsIds } from '../../src/types/ChainsIds';
 
 import { getTestContainer } from '../helpers/getTestContainer';
 
@@ -18,16 +18,17 @@ describe('ChainContractRepository', () => {
 
   describe('#get', () => {
     describe('Given a foreign chainId for an evm blockchain', async () => {
-      ForeignChainsIds.filter((x) => !NonEvmChainsIds.includes(x)).forEach((chainId) => {
-        it(`should return a ForeignChainContract instance for ${chainId}`, async () => {
-          const foreignChainContract = chainContractRepository.get(chainId);
-          await foreignChainContract.resolveContract();
-          expect(!!foreignChainContract).to.eql(true);
+      ForeignChainsIds.filter((x) => !NonEvmChainsIds.includes(x))
+        .filter((x) => x !== ChainsIds.ARBITRUM) // TODO enable when we move to goerli
+        .forEach((chainId) => {
+          it(`should return a ForeignChainContract instance for ${chainId}`, async () => {
+            const foreignChainContract = chainContractRepository.get(chainId);
+            expect(foreignChainContract).to.not.empty;
 
-          const address = foreignChainContract.address();
-          expect(typeof address).to.eql('string');
+            await foreignChainContract.resolveContract();
+            expect(foreignChainContract.address()).to.not.empty;
+          });
         });
-      });
     });
   });